requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.2
readability-lxml==0.8.1
openai==1.54.0
pydantic==2.10.2
pydantic-settings==2.6.1
certifi
